requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Video Processing
moviepy==1.0.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Audio Processing & Transcription
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# AI & Text-to-Speech
openai==1.6.1
edge-tts==6.1.9

# Utilities
numpy==1.26.2
requests==2.31.0
aiohttp==3.9.1
